# openapi/cpm-v1.yaml
openapi: 3.0.3
info:
  title: CAFE Crypto Policy Management (CPM) API
  version: "1.0.0"
  license:
    name: Proprietary
    url: https://example.com/license
  description: |
    Target HTTP contract for persisted policies, drafts, catalogue reads, and decision exploration
    per **WORKPLAN_API.md** §0.2, §2.3–2.4, §4.3–4.4, §5.1–5.2, §8.

    **Public URL prefix:** `/api/cpm/v1`. Do **not** publish paths like `/api/cpm/v1/cpm/...` (no duplicated `cpm` segment).

    **Discovery scans (correlation):** `cafe-discovery/openapi/discovery-v1.yaml` (PR1a); shared key **`scan_id`**.

    **Internal (service mesh / Docker, not public ingress):** gated by env
    **`CAFE_POLICY_REFERENCE_INTERNAL_SERVICE_TOKEN`** (Bearer).

    - `POST /internal/policies/references/scan` — request **`{ "scan_id", "user_id", "tenant_id" }`**;
      response **`{ "referenced": boolean, "count": number }`**. Discovery scan delete guard (WORKPLAN_API_PR **PR5**/**PR6**);
      shared lookup with **`GET …/policies?scan_id=`** (**PR7**).
    - `POST /internal/policies/references/wallet-target` — request **`{ "target_address", "user_id", "tenant_id" }`**
      (`target_address` normalized like Discovery wallet scans); response
      **`{ "exists": boolean, "policy_count": number, "draft_count": number }`**. Discovery **`POST …/scan`** W1 guard (**IMM-9b**).
servers:
  - url: https://{host}/api/cpm/v1
    description: Canonical product base (§0.2).
    variables:
      host:
        default: api.example.com

tags:
  - name: policies
    description: Persisted policy instances and related operations (§0.2).
  - name: drafts
    description: Policy drafts (§0.2).
  - name: catalog
    description: Static catalogue reads (graph, templates, reference instances).
  - name: decisions
    description: Non-persisted decision exploration (`explore`).

security:
  - bearerAuth: []

paths:
  /policies/catalog:
    get:
      tags: [catalog]
      summary: Get policy catalog (static)
      operationId: getPoliciesCatalog
      responses:
        "200":
          description: Catalogue payload (format as implemented; typically JSON or YAML-backed).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"

  /policies/templates:
    get:
      tags: [catalog]
      summary: List policy templates (static)
      operationId: getPoliciesTemplates
      responses:
        "200":
          description: Policy templates payload (static catalogue read).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"

  /policies/instances:
    get:
      tags: [catalog]
      summary: List reference / example instances from catalogue (static)
      operationId: getPoliciesInstancesCatalog
      responses:
        "200":
          description: Reference / example instances payload (static catalogue read).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"

  /policies/decisions/explore:
    post:
      tags: [decisions]
      summary: Explore applicable policies for a context (non-persistent)
      description: |
        **Does not** persist a final policy instance (WORKPLAN_API.md §5.1, §8.6). Body carries
        `policy_context` and `selection_request`. **`policy_context`** accepts either the flat Option A
        shape (`wallet_address` or **`target_address`**, `wallet_type`, `chain_ids`, …) or a Discovery v1
        wallet scan **detail** envelope (`scan_id`, `status`, **`result`** with `WalletScanResult` fields
        as in `cafe-discovery` `openapi/discovery-v1.yaml`). Discovery **`current_pq_posture`** labels
        `pq_ready` / `not_pq_ready` map to evaluator exports; `wallet_type` **`smart_account`** maps like
        Option A `AA`.
      operationId: postPoliciesDecisionsExplore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionExploreRequest"
      responses:
        "200":
          description: Evaluation result wrapper.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionExploreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /policies/assessment/request:
    post:
      tags: [policies]
      summary: Request asynchronous policy assessment for a Discovery wallet scan
      description: |
        **Wallet scans only** (WORKPLAN_API_PR PR13g). Replaces the removed Discovery route
        `POST /discovery/assessments/request`. TLS scans are **not** supported for CPM policy assessment.

        CPM fetches authoritative scan detail from Discovery (`GET /discovery/v1/wallets/scans/{scan_id}`)
        and emits `policy.assessment.requested.v0.1` on NATS. The request body must **not** include
        `policy_context` (server builds observation from Discovery).

        | Situation | HTTP |
        |-----------|------|
        | `policy_context` present | **400** |
        | Invalid `scan_id` / `selection_request` | **400** |
        | Authz deny / unknown scan / TLS scan / missing `result` | **404** |
        | Discovery authz or detail unavailable | **503** |
        | Accepted for async processing | **202** |
      operationId: postPoliciesAssessmentRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyAssessmentRequest"
      responses:
        "202":
          description: Assessment command accepted for async processing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyAssessmentRequestAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Scan not found or not eligible (owner-scoped wallet scan only).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Discovery or NATS publish unavailable (fail-closed).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /policies:
    get:
      tags: [policies]
      summary: Get persisted policy instance(s)
      description: |
        **Exactly one** of **`id`** (single instance) or **`scan_id`** (list owner-scoped) must be used.
        **Both `id` and `scan_id` → 400** (WORKPLAN_API.md §5.2, §8.2).
      operationId: getPolicies
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: string
          description: Return a single persisted instance by id.
        - name: scan_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: List instances referencing this scan (owner-scoped).
      responses:
        "200":
          description: Single `PolicyRecord` or `PolicyListByScanIdResponse` depending on query mode.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PolicyRecord"
                  - $ref: "#/components/schemas/PolicyListByScanIdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown id (single-instance mode only).
    post:
      tags: [policies]
      summary: Persist a policy instance
      description: |
        **`scan_id` is mandatory** for instances created from Discovery-bound flows (WORKPLAN_API.md §5.2).
        May be absent only for catalogue fixtures, manual drafts not yet bound, or explicit non-product tests
        — distinguish in validation and tests.
      operationId: postPolicies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyRecordWrite"
      responses:
        "200":
          description: Created or updated instance as per implementation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [policies]
      summary: Delete persisted policy instance by id
      description: |
        **204** if deleted; **404** if unknown / out of scope / already deleted (idempotent). **Never**
        deletes the Discovery scan row (WORKPLAN_API.md §5.2, §8.5). **No 409** on this route in this plan.
      operationId: deletePolicies
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Instance removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /drafts:
    get:
      tags: [drafts]
      summary: List drafts (owner)
      operationId: getDrafts
      responses:
        "200":
          description: Owner-scoped list of drafts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyDraft"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [drafts]
      summary: Create or update a draft
      operationId: postDrafts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDraftWrite"
      responses:
        "200":
          description: Draft created or updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [drafts]
      summary: Delete platform draft by id
      description: |
        **204** if deleted; **404** if unknown / out of scope / already deleted (idempotent). Removes
        the platform draft to satisfy **W1** and unblock **`POST /api/discovery/v1/scan`** (WORKPLAN_API.md
        §2.2, §2.4). **Never** deletes the Discovery scan row referenced by the draft.
      operationId: deleteDrafts
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Draft removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Unauthorized:
      description: Missing or invalid JWT.
    BadRequest:
      description: Invalid input (including `id` + `scan_id` together on GET `/policies`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Unknown resource or idempotent delete miss.

  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    PolicyRecord:
      type: object
      required: [id, payload]
      description: Persisted instance (WORKPLAN_API.md §4.4, §8.3 `PolicyRecord`).
      properties:
        id:
          type: string
        scan_id:
          type: string
          format: uuid
          description: Required for Discovery→CPM flows; optional only for catalogue/fixture/manual-exception cases (§5.2).
        payload:
          type: object
          additionalProperties: true

    PolicyRecordWrite:
      type: object
      required: [id, payload]
      properties:
        id:
          type: string
        scan_id:
          type: string
          format: uuid
        binding:
          type: string
          enum: [discovery, fixture, catalog, none]
          description: |
            **`discovery`** (default when omitted): `scan_id` is mandatory (UUID) for Discovery→CPM persisted instances.
            **`fixture`**, **`catalog`**, **`none`**: optional `scan_id` for non-product tests and catalogue-aligned payloads (WORKPLAN_API.md §5.2).
        payload:
          type: object
          additionalProperties: true

    PolicyListByScanIdResponse:
      type: object
      required: [items, total, limit, offset]
      description: Owner-scoped list for `GET …/policies?scan_id=` (WORKPLAN_API.md §5.2, §8.3).
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PolicyRecord"
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0

    PolicyDraft:
      type: object
      additionalProperties: true

    PolicyDraftWrite:
      type: object
      additionalProperties: true

    DecisionExploreRequest:
      type: object
      required: [policy_context, selection_request]
      properties:
        scan_id:
          type: string
          description: Optional scan binding for AUTH-02 (implementation detail).
        policy_context:
          type: object
          description: |
            Flat Option A context and/or Discovery v1 `WalletScanDetail`-shaped object (`result` for terminal scans).
            Extra keys are ignored when parsing the nested `result` branch; unknown keys at the top level of the
            request body remain rejected by the API (strict JSON on the outer envelope).
          additionalProperties: true
        selection_request:
          type: object
          additionalProperties: true

    DecisionExploreResponse:
      type: object
      required: [decision]
      properties:
        decision:
          type: object
          additionalProperties: true

    PolicyAssessmentRequest:
      type: object
      required: [scan_id, selection_request]
      description: |
        Async policy assessment trigger for a **Discovery wallet scan** only. Do **not** send
        `policy_context` (returns **400** if present). CPM reconstructs observation from Discovery
        wallet scan detail (`GET /discovery/v1/wallets/scans/{scan_id}`).
      properties:
        scan_id:
          type: string
          format: uuid
          description: Discovery wallet scan id (owner-scoped).
        selection_request:
          type: object
          description: Same shape as `selection_request` on `POST …/decisions/explore`.
          additionalProperties: true
        client_request_id:
          type: string
          description: Optional client correlation id (not the NATS idempotence key).
      additionalProperties: false

    PolicyAssessmentRequestAccepted:
      type: object
      required: [event_id, correlation_id]
      properties:
        event_id:
          type: string
          description: Idempotence key for `policy.assessment.requested.v0.1`.
        correlation_id:
          type: string
          description: Typically the `scan_id`.
        client_request_id:
          type: string
          description: Echo of optional client request id when provided.